+++ resources/views/components/breadcrumb/ellipsis.blade.php
@@ -1,7 +1,7 @@
 @php
 use DistortedFusion\BladeComponents\BladeComponents;
 @endphp
-<li {{ $attributes->class(['leading-none']) }}>
+<li data-slot="breadcrumb-ellipsis" {{ $attributes->class(['leading-none']) }}>
     <x-dynamic-component :component="BladeComponents::defaultBreadcrumbEllipsisIcon()"
         class="size-4 text-[var(--muted-foreground)]" />
 </li>