+++ resources/views/components/breadcrumb/index.blade.php
@@ -1,4 +1,4 @@
-<ol {{ $attributes->class([
+<ol data-slot="breadcrumb" {{ $attributes->class([
     'flex items-center gap-x-1 md:gap-x-2',
 ]) }}>
     {{ $slot }}