+++ resources/views/components/breadcrumb/separator.blade.php
@@ -1,7 +1,7 @@
 @php
 use DistortedFusion\BladeComponents\BladeComponents;
 @endphp
-<li {{ $attributes->class(['leading-none']) }}>
+<li data-slot="breadcrumb-separator" {{ $attributes->class(['leading-none']) }}>
     <x-dynamic-component :component="BladeComponents::defaultBreadcrumbSeparatorIcon()"
         class="size-3 text-[var(--foreground)]" />
 </li>